+++ modules/sprintf.js
@@ -1,21 +1,22 @@
-sprintf  = require('sprintf-js').sprintf
-vsprintf = require('sprintf-js').vsprintf
-traverse = require('./traverse')
+const { sprintf }  = require('sprintf-js')
+const { vsprintf } = require('sprintf-js')
+const traverse = require('./traverse')
 
-###*
+/**
  * Replace placeholders with provided values via `sprintf` or `vsprintf`.
  * `printf` type will be chosen depending on provided placeholders
  * @param {*}                   input           Input in which should be made replacement
  * @param {string|object|array} placeholders... List of placeholders, object with named
  *                                              placeholders or arrays of placeholders
  * @return {*} Transformed input
-###
-module.exports = (input, placeholders...) ->
-  [placeholder, ...] = placeholders
-  useVsprintf = placeholders.length == 1 and Array.isArray(placeholder)
+*/
+module.exports = (input, ...placeholders) => {
+  const [placeholder] = placeholders
+  const useVsprintf = placeholders.length === 1 && Array.isArray(placeholder)
 
   return traverse(
     input,
-    (tmpl) => if useVsprintf then vsprintf tmpl, placeholder else sprintf tmpl, placeholders...
+    (tmpl) => useVsprintf ? vsprintf(tmpl, placeholder) : sprintf(tmpl, ...placeholders),
     (string) => /%[^\s]/.test(string)
   )
+}